misc/HtmlTestSuite/results/test-0063.xhtml

<?xml version="1.0" encoding="UTF-8"?><html xmlns="http://www.w3.org/1999/xhtml">
<head>
<meta charset="utf-8"/>
<title>Dynamic childlist of MathML elements</title>
<script src="/mathml/support/mathml-fragments.js"/>
<link href="https://w3c.github.io/mathml-core/#adjust-space-around-content-mpadded" rel="help"/>
<link href="https://w3c.github.io/mathml-core/#dom-and-javascript" rel="help"/>
<link href="https://w3c.github.io/mathml-core/#fractions-mfrac" rel="help"/>
<link href="https://w3c.github.io/mathml-core/#prescripts-and-tensor-indices-mmultiscripts" rel="help"/>
<link href="https://w3c.github.io/mathml-core/#radicals-msqrt-mroot" rel="help"/>
<link href="https://w3c.github.io/mathml-core/#space-mspace" rel="help"/>
<link href="https://w3c.github.io/mathml-core/#subscripts-and-superscripts-msub-msup-msubsup" rel="help"/>
<link href="https://w3c.github.io/mathml-core/#underscripts-and-overscripts-munder-mover-munderover" rel="help"/>
<meta content="Dynamically modify DOM tree of some MathML elements by adding or removing children." name="assert"/>
<script src="/resources/testharness.js"/>
<script src="/resources/testharnessreport.js"/>
<script src="/mathml/support/feature-detection.js"/>
<script src="/mathml/support/layout-comparison.js"/>
<script>
  function forceNumberOfChildren(element, count) {
      while (element.children.length &gt; count)
          element.removeChild(element.lastElementChild);
      for (let i = element.children.length; i &lt; count; i++) {
          if (element.tagName === "mmultiscripts" &amp;&amp; i === 5) {
              element.appendChild(FragmentHelper.createElement("mprescripts"));
          } else {
              let mspace = FragmentHelper.createElement("mspace");
              mspace.setAttribute("width", `10px`);
              mspace.setAttribute("height", `${10*(i+1)}px`);
              mspace.setAttribute("style", `background: black;`);
              element.appendChild(mspace);
          }
      }
  }

  setup({ explicit_done: true });
  window.addEventListener("load", function() {
      // force initial layout so we're sure what we're testing against
      document.documentElement.getBoundingClientRect();

      let reference = document.getElementById("reference");

      Array.from(document.querySelectorAll("[data-title]")).forEach(element =&gt; {
          test(function() {
              assert_true(MathMLFeatureDetection.has_mspace());
              let reference = document.getElementById(`${element.getAttribute("data-reference")}`);
              forceNumberOfChildren(element, reference.children.length);
              const epsilon = 1;
              if (element.tagName == "mspace") {
                  compareSize(element, reference, epsilon);
                  childrenHaveEmptyBoundingClientRects(element);
                  childrenHaveEmptyBoundingClientRects(reference);
              } else {
                  compareLayout(element, reference, epsilon);
              }
          }, `${element.getAttribute("data-title")}`);
      });
      done();
  });
</script>
</head>
<body>
  <div id="log"/>
  <p>
    <math xmlns="http://www.w3.org/1998/Math/MathML">
      <mfrac id="mfrac-reference-1">
        <mspace height="10px" style="background: black;" width="10px"/>
      </mfrac>
      <mfrac id="mfrac-reference-2">
        <mspace height="10px" style="background: black;" width="10px"/>
        <mspace height="20px" style="background: black;" width="10px"/>
      </mfrac>
      <mfrac id="mfrac-reference-3">
        <mspace height="10px" style="background: black;" width="10px"/>
        <mspace height="20px" style="background: black;" width="10px"/>
        <mspace height="30px" style="background: black;" width="10px"/>
      </mfrac>
    </math>
  </p>
  <p>
    <math xmlns="http://www.w3.org/1998/Math/MathML">
      <mfrac data-reference="mfrac-reference-2" data-title="Adding missing children to mfrac">
      </mfrac>
    </math>
  </p>
  <p>
    <math xmlns="http://www.w3.org/1998/Math/MathML">
      <mfrac data-reference="mfrac-reference-1" data-title="Removing child from valid mfrac">
        <mspace height="10px" style="background: black;" width="10px"/>
        <mspace height="20px" style="background: black;" width="10px"/>
      </mfrac>
    </math>
  </p>
  <p>
    <math xmlns="http://www.w3.org/1998/Math/MathML">
      <mfrac data-reference="mfrac-reference-3" data-title="Adding child to valid mfrac">
        <mspace height="10px" style="background: black;" width="10px"/>
        <mspace height="20px" style="background: black;" width="10px"/>
      </mfrac>
    </math>
  </p>
  <p>
    <math xmlns="http://www.w3.org/1998/Math/MathML">
      <mfrac data-reference="mfrac-reference-2" data-title="Removing extra child from mfrac">
        <mspace height="10px" style="background: black;" width="10px"/>
        <mspace height="20px" style="background: black;" width="10px"/>
        <mspace height="30px" style="background: black;" width="10px"/>
      </mfrac>
    </math>
  </p>
  <hr/>
  <p>
    <math xmlns="http://www.w3.org/1998/Math/MathML">
      <munder id="munder-reference-1">
        <mspace height="10px" style="background: black;" width="10px"/>
      </munder>
      <munder id="munder-reference-2">
        <mspace height="10px" style="background: black;" width="10px"/>
        <mspace height="20px" style="background: black;" width="10px"/>
      </munder>
      <munder id="munder-reference-3">
        <mspace height="10px" style="background: black;" width="10px"/>
        <mspace height="20px" style="background: black;" width="10px"/>
        <mspace height="30px" style="background: black;" width="10px"/>
      </munder>
    </math>
  </p>
  <p>
    <math xmlns="http://www.w3.org/1998/Math/MathML">
      <munder data-reference="munder-reference-2" data-title="Adding missing children to munder">
      </munder>
    </math>
  </p>
  <p>
    <math xmlns="http://www.w3.org/1998/Math/MathML">
      <munder data-reference="munder-reference-1" data-title="Removing child from valid munder">
        <mspace height="10px" style="background: black;" width="10px"/>
        <mspace height="20px" style="background: black;" width="10px"/>
      </munder>
    </math>
  </p>
  <p>
    <math xmlns="http://www.w3.org/1998/Math/MathML">
      <munder data-reference="munder-reference-3" data-title="Adding child to valid munder">
        <mspace height="10px" style="background: black;" width="10px"/>
        <mspace height="20px" style="background: black;" width="10px"/>
      </munder>
    </math>
  </p>
  <p>
    <math xmlns="http://www.w3.org/1998/Math/MathML">
      <munder data-reference="munder-reference-2" data-title="Removing extra child from munder">
        <mspace height="10px" style="background: black;" width="10px"/>
        <mspace height="20px" style="background: black;" width="10px"/>
        <mspace height="30px" style="background: black;" width="10px"/>
      </munder>
    </math>
  </p>
  <hr/>
  <p>
    <math xmlns="http://www.w3.org/1998/Math/MathML">
      <mover id="mover-reference-1">
        <mspace height="10px" style="background: black;" width="10px"/>
      </mover>
      <mover id="mover-reference-2">
        <mspace height="10px" style="background: black;" width="10px"/>
        <mspace height="20px" style="background: black;" width="10px"/>
      </mover>
      <mover id="mover-reference-3">
        <mspace height="10px" style="background: black;" width="10px"/>
        <mspace height="20px" style="background: black;" width="10px"/>
        <mspace height="30px" style="background: black;" width="10px"/>
      </mover>
    </math>
  </p>
  <p>
    <math xmlns="http://www.w3.org/1998/Math/MathML">
      <mover data-reference="mover-reference-2" data-title="Adding missing children to mover">
      </mover>
    </math>
  </p>
  <p>
    <math xmlns="http://www.w3.org/1998/Math/MathML">
      <mover data-reference="mover-reference-1" data-title="Removing child from valid mover">
        <mspace height="10px" style="background: black;" width="10px"/>
        <mspace height="20px" style="background: black;" width="10px"/>
      </mover>
    </math>
  </p>
  <p>
    <math xmlns="http://www.w3.org/1998/Math/MathML">
      <mover data-reference="mover-reference-3" data-title="Adding child to valid mover">
        <mspace height="10px" style="background: black;" width="10px"/>
        <mspace height="20px" style="background: black;" width="10px"/>
      </mover>
    </math>
  </p>
  <p>
    <math xmlns="http://www.w3.org/1998/Math/MathML">
      <mover data-reference="mover-reference-2" data-title="Removing extra child from mover">
        <mspace height="10px" style="background: black;" width="10px"/>
        <mspace height="20px" style="background: black;" width="10px"/>
        <mspace height="30px" style="background: black;" width="10px"/>
      </mover>
    </math>
  </p>
  <hr/>
  <p>
    <math xmlns="http://www.w3.org/1998/Math/MathML">
      <munderover id="munderover-reference-2">
        <mspace height="10px" style="background: black;" width="10px"/>
        <mspace height="20px" style="background: black;" width="10px"/>
      </munderover>
      <munderover id="munderover-reference-3">
        <mspace height="10px" style="background: black;" width="10px"/>
        <mspace height="20px" style="background: black;" width="10px"/>
        <mspace height="30px" style="background: black;" width="10px"/>
      </munderover>
      <munderover id="munderover-reference-4">
        <mspace height="10px" style="background: black;" width="10px"/>
        <mspace height="20px" style="background: black;" width="10px"/>
        <mspace height="30px" style="background: black;" width="10px"/>
        <mspace height="40px" style="background: black;" width="10px"/>
      </munderover>
    </math>
  </p>
  <p>
    <math xmlns="http://www.w3.org/1998/Math/MathML">
      <munderover data-reference="munderover-reference-3" data-title="Adding missing children to munderover">
      </munderover>
    </math>
  </p>
  <p>
    <math xmlns="http://www.w3.org/1998/Math/MathML">
      <munderover data-reference="munderover-reference-2" data-title="Removing child from valid munderover">
        <mspace height="10px" style="background: black;" width="10px"/>
        <mspace height="20px" style="background: black;" width="10px"/>
        <mspace height="30px" style="background: black;" width="10px"/>
      </munderover>
    </math>
  </p>
  <p>
    <math xmlns="http://www.w3.org/1998/Math/MathML">
      <munderover data-reference="munderover-reference-4" data-title="Adding child to valid munderover">
        <mspace height="10px" style="background: black;" width="10px"/>
        <mspace height="20px" style="background: black;" width="10px"/>
        <mspace height="30px" style="background: black;" width="10px"/>
      </munderover>
    </math>
  </p>
  <p>
    <math xmlns="http://www.w3.org/1998/Math/MathML">
      <munderover data-reference="munderover-reference-3" data-title="Removing extra child from munderover">
        <mspace height="10px" style="background: black;" width="10px"/>
        <mspace height="20px" style="background: black;" width="10px"/>
        <mspace height="30px" style="background: black;" width="10px"/>
        <mspace height="40px" style="background: black;" width="10px"/>
      </munderover>
    </math>
  </p>
  <hr/>
  <p>
    <math xmlns="http://www.w3.org/1998/Math/MathML">
      <msub id="msub-reference-1">
        <mspace height="10px" style="background: black;" width="10px"/>
      </msub>
      <msub id="msub-reference-2">
        <mspace height="10px" style="background: black;" width="10px"/>
        <mspace height="20px" style="background: black;" width="10px"/>
      </msub>
      <msub id="msub-reference-3">
        <mspace height="10px" style="background: black;" width="10px"/>
        <mspace height="20px" style="background: black;" width="10px"/>
        <mspace height="30px" style="background: black;" width="10px"/>
      </msub>
    </math>
  </p>
  <p>
    <math xmlns="http://www.w3.org/1998/Math/MathML">
      <msub data-reference="msub-reference-2" data-title="Adding missing children to msub">
      </msub>
    </math>
  </p>
  <p>
    <math xmlns="http://www.w3.org/1998/Math/MathML">
      <msub data-reference="msub-reference-1" data-title="Removing child from valid msub">
        <mspace height="10px" style="background: black;" width="10px"/>
        <mspace height="20px" style="background: black;" width="10px"/>
      </msub>
    </math>
  </p>
  <p>
    <math xmlns="http://www.w3.org/1998/Math/MathML">
      <msub data-reference="msub-reference-3" data-title="Adding child to valid msub">
        <mspace height="10px" style="background: black;" width="10px"/>
        <mspace height="20px" style="background: black;" width="10px"/>
      </msub>
    </math>
  </p>
  <p>
    <math xmlns="http://www.w3.org/1998/Math/MathML">
      <msub data-reference="msub-reference-2" data-title="Removing extra child from msub">
        <mspace height="10px" style="background: black;" width="10px"/>
        <mspace height="20px" style="background: black;" width="10px"/>
        <mspace height="30px" style="background: black;" width="10px"/>
      </msub>
    </math>
  </p>
  <hr/>
  <p>
    <math xmlns="http://www.w3.org/1998/Math/MathML">
      <msup id="msup-reference-1">
        <mspace height="10px" style="background: black;" width="10px"/>
      </msup>
      <msup id="msup-reference-2">
        <mspace height="10px" style="background: black;" width="10px"/>
        <mspace height="20px" style="background: black;" width="10px"/>
      </msup>
      <msup id="msup-reference-3">
        <mspace height="10px" style="background: black;" width="10px"/>
        <mspace height="20px" style="background: black;" width="10px"/>
        <mspace height="30px" style="background: black;" width="10px"/>
      </msup>
    </math>
  </p>
  <p>
    <math xmlns="http://www.w3.org/1998/Math/MathML">
      <msup data-reference="msup-reference-2" data-title="Adding missing children to msup">
      </msup>
    </math>
  </p>
  <p>
    <math xmlns="http://www.w3.org/1998/Math/MathML">
      <msup data-reference="msup-reference-1" data-title="Removing child from valid msup">
        <mspace height="10px" style="background: black;" width="10px"/>
        <mspace height="20px" style="background: black;" width="10px"/>
      </msup>
    </math>
  </p>
  <p>
    <math xmlns="http://www.w3.org/1998/Math/MathML">
      <msup data-reference="msup-reference-3" data-title="Adding child to valid msup">
        <mspace height="10px" style="background: black;" width="10px"/>
        <mspace height="20px" style="background: black;" width="10px"/>
      </msup>
    </math>
  </p>
  <p>
    <math xmlns="http://www.w3.org/1998/Math/MathML">
      <msup data-reference="msup-reference-2" data-title="Removing extra child from msup">
        <mspace height="10px" style="background: black;" width="10px"/>
        <mspace height="20px" style="background: black;" width="10px"/>
        <mspace height="30px" style="background: black;" width="10px"/>
      </msup>
    </math>
  </p>
  <hr/>
  <p>
    <math xmlns="http://www.w3.org/1998/Math/MathML">
      <msubsup id="msubsup-reference-2">
        <mspace height="10px" style="background: black;" width="10px"/>
        <mspace height="20px" style="background: black;" width="10px"/>
      </msubsup>
      <msubsup id="msubsup-reference-3">
        <mspace height="10px" style="background: black;" width="10px"/>
        <mspace height="20px" style="background: black;" width="10px"/>
        <mspace height="30px" style="background: black;" width="10px"/>
      </msubsup>
      <msubsup id="msubsup-reference-4">
        <mspace height="10px" style="background: black;" width="10px"/>
        <mspace height="20px" style="background: black;" width="10px"/>
        <mspace height="30px" style="background: black;" width="10px"/>
        <mspace height="40px" style="background: black;" width="10px"/>
      </msubsup>
    </math>
  </p>
  <p>
    <math xmlns="http://www.w3.org/1998/Math/MathML">
      <msubsup data-reference="msubsup-reference-3" data-title="Adding missing children to msubsup">
      </msubsup>
    </math>
  </p>
  <p>
    <math xmlns="http://www.w3.org/1998/Math/MathML">
      <msubsup data-reference="msubsup-reference-2" data-title="Removing child from valid msubsup">
        <mspace height="10px" style="background: black;" width="10px"/>
        <mspace height="20px" style="background: black;" width="10px"/>
        <mspace height="30px" style="background: black;" width="10px"/>
      </msubsup>
    </math>
  </p>
  <p>
    <math xmlns="http://www.w3.org/1998/Math/MathML">
      <msubsup data-reference="msubsup-reference-4" data-title="Adding child to valid msubsup">
        <mspace height="10px" style="background: black;" width="10px"/>
        <mspace height="20px" style="background: black;" width="10px"/>
        <mspace height="30px" style="background: black;" width="10px"/>
      </msubsup>
    </math>
  </p>
  <p>
    <math xmlns="http://www.w3.org/1998/Math/MathML">
      <msubsup data-reference="msubsup-reference-3" data-title="Removing extra child from msubsup">
        <mspace height="10px" style="background: black;" width="10px"/>
        <mspace height="20px" style="background: black;" width="10px"/>
        <mspace height="30px" style="background: black;" width="10px"/>
        <mspace height="40px" style="background: black;" width="10px"/>
      </msubsup>
    </math>
  </p>
  <hr/>
  <p>
    <math xmlns="http://www.w3.org/1998/Math/MathML">
      <mroot id="mroot-reference-1">
        <mspace height="10px" style="background: black;" width="10px"/>
      </mroot>
      <mroot id="mroot-reference-2">
        <mspace height="10px" style="background: black;" width="10px"/>
        <mspace height="20px" style="background: black;" width="10px"/>
      </mroot>
      <mroot id="mroot-reference-3">
        <mspace height="10px" style="background: black;" width="10px"/>
        <mspace height="20px" style="background: black;" width="10px"/>
        <mspace height="30px" style="background: black;" width="10px"/>
      </mroot>
    </math>
  </p>
  <p>
    <math xmlns="http://www.w3.org/1998/Math/MathML">
      <mroot data-reference="mroot-reference-2" data-title="Adding missing children to mroot">
      </mroot>
    </math>
  </p>
  <p>
    <math xmlns="http://www.w3.org/1998/Math/MathML">
      <mroot data-reference="mroot-reference-1" data-title="Removing child from valid mroot">
        <mspace height="10px" style="background: black;" width="10px"/>
        <mspace height="20px" style="background: black;" width="10px"/>
      </mroot>
    </math>
  </p>
  <p>
    <math xmlns="http://www.w3.org/1998/Math/MathML">
      <mroot data-reference="mroot-reference-3" data-title="Adding child to valid mroot">
        <mspace height="10px" style="background: black;" width="10px"/>
        <mspace height="20px" style="background: black;" width="10px"/>
      </mroot>
    </math>
  </p>
  <p>
    <math xmlns="http://www.w3.org/1998/Math/MathML">
      <mroot data-reference="mroot-reference-2" data-title="Removing extra child from mroot">
        <mspace height="10px" style="background: black;" width="10px"/>
        <mspace height="20px" style="background: black;" width="10px"/>
        <mspace height="30px" style="background: black;" width="10px"/>
      </mroot>
    </math>
  </p>
  <hr/>
  <p>
    <math xmlns="http://www.w3.org/1998/Math/MathML">
      <msqrt id="msqrt-reference-0">
      </msqrt>
      <msqrt id="msqrt-reference-2">
        <mspace height="10px" style="background: black;" width="10px"/>
        <mspace height="20px" style="background: black;" width="10px"/>
       </msqrt>
    </math>
  </p>
  <p>
    <math xmlns="http://www.w3.org/1998/Math/MathML">
      <msqrt data-reference="msqrt-reference-0" data-title="Removing children from msqrt">
        <mspace height="10px" style="background: black;" width="10px"/>
        <mspace height="20px" style="background: black;" width="10px"/>
        <mspace height="30px" style="background: black;" width="10px"/>
      </msqrt>
    </math>
  </p>
  <p>
    <math xmlns="http://www.w3.org/1998/Math/MathML">
      <msqrt data-reference="msqrt-reference-2" data-title="Adding children to msqrt">
        <mspace height="10px" style="background: black;" width="10px"/>
      </msqrt>
    </math>
  </p>
  <hr/>
  <p>
    <math xmlns="http://www.w3.org/1998/Math/MathML">
      <mpadded id="mpadded-reference-0">
      </mpadded>
      <mpadded id="mpadded-reference-2">
        <mspace height="10px" style="background: black;" width="10px"/>
        <mspace height="20px" style="background: black;" width="10px"/>
       </mpadded>
    </math>
  </p>
  <p>
    <math xmlns="http://www.w3.org/1998/Math/MathML">
      <mpadded data-reference="mpadded-reference-0" data-title="Removing children from mpadded">
        <mspace height="10px" style="background: black;" width="10px"/>
        <mspace height="20px" style="background: black;" width="10px"/>
        <mspace height="30px" style="background: black;" width="10px"/>
      </mpadded>
    </math>
  </p>
  <p>
    <math xmlns="http://www.w3.org/1998/Math/MathML">
      <mpadded data-reference="mpadded-reference-2" data-title="Adding children to mpadded">
        <mspace height="10px" style="background: black;" width="10px"/>
      </mpadded>
    </math>
  </p>
  <hr/>
  <p>
    <math xmlns="http://www.w3.org/1998/Math/MathML">
      <mspace height="70px" id="mspace-reference-0" style="background: blue" width="30px">
      </mspace>
      <mspace height="70px" id="mspace-reference-2" style="background: green" width="30px">
        <mspace height="10px" style="background: black;" width="10px"/>
        <mspace height="20px" style="background: black;" width="10px"/>
       </mspace>
    </math>
  </p>
  <p>
    <math xmlns="http://www.w3.org/1998/Math/MathML">
      <mspace data-reference="mspace-reference-0" data-title="Removing children from mspace" height="70px" style="background: lightblue" width="30px">
        <mspace height="10px" style="background: black;" width="10px"/>
        <mspace height="20px" style="background: black;" width="10px"/>
        <mspace height="30px" style="background: black;" width="10px"/>
      </mspace>
    </math>
  </p>
  <p>
    <math xmlns="http://www.w3.org/1998/Math/MathML">
      <mspace data-reference="mspace-reference-2" data-title="Adding children to mspace" height="70px" style="background: lightgreen" width="30px">
        <mspace height="10px" style="background: black;" width="10px"/>
      </mspace>
    </math>
  </p>
  <hr/>
  <p>
    <math xmlns="http://www.w3.org/1998/Math/MathML">
      <mmultiscripts id="mmultiscripts-reference-0">
      </mmultiscripts>
      <mmultiscripts id="mmultiscripts-reference-1">
        <mspace height="10px" style="background: black;" width="10px"/>
      </mmultiscripts>
      <mmultiscripts id="mmultiscripts-reference-2">
        <mspace height="10px" style="background: black;" width="10px"/>
        <mspace height="20px" style="background: black;" width="10px"/>
      </mmultiscripts>
      <mmultiscripts id="mmultiscripts-reference-3">
        <mspace height="10px" style="background: black;" width="10px"/>
        <mspace height="20px" style="background: black;" width="10px"/>
        <mspace height="30px" style="background: black;" width="10px"/>
      </mmultiscripts>
      <mmultiscripts id="mmultiscripts-reference-6">
        <mspace height="10px" style="background: black;" width="10px"/>
        <mspace height="20px" style="background: black;" width="10px"/>
        <mspace height="30px" style="background: black;" width="10px"/>
        <mspace height="40px" style="background: black;" width="10px"/>
        <mspace height="50px" style="background: black;" width="10px"/>
        <mprescripts/>
      </mmultiscripts>
      <mmultiscripts id="mmultiscripts-reference-8">
        <mspace height="10px" style="background: black;" width="10px"/>
        <mspace height="20px" style="background: black;" width="10px"/>
        <mspace height="30px" style="background: black;" width="10px"/>
        <mspace height="40px" style="background: black;" width="10px"/>
        <mspace height="50px" style="background: black;" width="10px"/>
        <mprescripts/>
        <mspace height="70px" style="background: black;" width="10px"/>
        <mspace height="80px" style="background: black;" width="10px"/>
      </mmultiscripts>
    </math>
  </p>
  <p>
    <math xmlns="http://www.w3.org/1998/Math/MathML">
      <mmultiscripts data-reference="mmultiscripts-reference-0" data-title="multiscripts child count from 3 to 0">
        <mspace height="10px" style="background: black;" width="10px"/>
        <mspace height="20px" style="background: black;" width="10px"/>
        <mspace height="30px" style="background: black;" width="10px"/>
      </mmultiscripts>
    </math>
  </p>
  <p>
    <math xmlns="http://www.w3.org/1998/Math/MathML">
      <mmultiscripts data-reference="mmultiscripts-reference-1" data-title="multiscripts child count from 3 to 1">
        <mspace height="10px" style="background: black;" width="10px"/>
        <mspace height="20px" style="background: black;" width="10px"/>
        <mspace height="30px" style="background: black;" width="10px"/>
      </mmultiscripts>
    </math>
  </p>
  <p>
    <math xmlns="http://www.w3.org/1998/Math/MathML">
      <mmultiscripts data-reference="mmultiscripts-reference-2" data-title="multiscripts child count from 3 to 2">
        <mspace height="10px" style="background: black;" width="10px"/>
        <mspace height="20px" style="background: black;" width="10px"/>
        <mspace height="30px" style="background: black;" width="10px"/>
      </mmultiscripts>
    </math>
  </p>
  <p>
    <math xmlns="http://www.w3.org/1998/Math/MathML">
      <mmultiscripts data-reference="mmultiscripts-reference-3" data-title="multiscripts child count from 0 to 3">
      </mmultiscripts>
    </math>
  </p>
  <p>
    <math xmlns="http://www.w3.org/1998/Math/MathML">
      <mmultiscripts data-reference="mmultiscripts-reference-6" data-title="multiscripts child count from 3 to 6">
        <mspace height="10px" style="background: black;" width="10px"/>
        <mspace height="20px" style="background: black;" width="10px"/>
        <mspace height="30px" style="background: black;" width="10px"/>
      </mmultiscripts>
    </math>
  </p>
  <p>
    <math xmlns="http://www.w3.org/1998/Math/MathML">
      <mmultiscripts data-reference="mmultiscripts-reference-8" data-title="multiscripts child count from 3 to 8">
        <mspace height="10px" style="background: black;" width="10px"/>
        <mspace height="20px" style="background: black;" width="10px"/>
        <mspace height="30px" style="background: black;" width="10px"/>
      </mmultiscripts>
    </math>
  </p>
</body>
</html>
